+++ reactos/lib/sdk/crt/wine/undname.c
@@ -53,7 +53,7 @@ WINE_DEFAULT_DEBUG_CHANNEL(msvcrt);
 #define UNDNAME_NO_COMPLEX_TYPE          (0x8000)
 
 /* How data types modifiers are stored:
- * M (in the following definitions) is defined for 
+ * M (in the following definitions) is defined for
  * 'A', 'B', 'C' and 'D' as follows
  *      {<A>}:  ""
  *      {<B>}:  "const "
@@ -67,7 +67,7 @@ WINE_DEFAULT_DEBUG_CHANNEL(msvcrt);
  *      in data fields:
  *              same as for arguments and also the following
  *              ?<M>x   {<M>}x
- *              
+ *
  */
 
 struct array
@@ -127,7 +127,7 @@ static void*    und_alloc(struct parsed_symbol* sym, unsigned int len)
         sym->avail_in_first = 0;
         ptr = (char*)sym->alloc_list + sizeof(void*);
     }
-    else 
+    else
     {
         if (len > sym->avail_in_first)
         {
@@ -201,14 +201,14 @@ static BOOL str_array_push(struct parsed_symbol* sym, const char* ptr, int len,
         a->alloc *= 2;
         a->elts = new;
     }
-    if (len == -1) len = strlen(ptr);
+    if (len == -1) len = (int)strlen(ptr);
     a->elts[a->num] = und_alloc(sym, len + 1);
     assert(a->elts[a->num]);
     memcpy(a->elts[a->num], ptr, len);
-    a->elts[a->num][len] = '\0'; 
+    a->elts[a->num][len] = '\0';
     if (++a->num >= a->max) a->max = a->num;
     {
-        int i;
+        unsigned int i;
         char c;
 
         for (i = a->max - 1; i >= 0; i--)
@@ -234,18 +234,18 @@ static char* str_array_get_ref(struct array* cref, unsigned idx)
     assert(cref);
     if (cref->start + idx >= cref->max)
     {
-        WARN("Out of bounds: %p %d + %d >= %d\n", 
+        WARN("Out of bounds: %p %d + %d >= %d\n",
               cref, cref->start, idx, cref->max);
         return NULL;
     }
-    TRACE("Returning %p[%d] => %s\n", 
+    TRACE("Returning %p[%d] => %s\n",
           cref, idx, cref->elts[cref->start + idx]);
     return cref->elts[cref->start + idx];
 }
 
 /******************************************************************
  *		str_printf
- * Helper for printf type of command (only %s and %c are implemented) 
+ * Helper for printf type of command (only %s and %c are implemented)
  * while dynamically allocating the buffer
  */
 static char* str_printf(struct parsed_symbol* sym, const char* format, ...)
@@ -263,7 +263,7 @@ static char* str_printf(struct parsed_symbol* sym, const char* format, ...)
         {
             switch (format[++i])
             {
-            case 's': t = va_arg(args, char*); if (t) len += strlen(t); break;
+            case 's': t = va_arg(args, char*); if (t) len += (int)strlen(t); break;
             case 'c': (void)va_arg(args, int); len++; break;
             default: i--; /* fall thru */
             case '%': len++; break;
@@ -284,7 +284,7 @@ static char* str_printf(struct parsed_symbol* sym, const char* format, ...)
                 t = va_arg(args, char*);
                 if (t)
                 {
-                    sz = strlen(t);
+                    sz = (int)strlen(t);
                     memcpy(p, t, sz);
                     p += sz;
                 }
@@ -358,7 +358,7 @@ static const char* get_number(struct parsed_symbol* sym)
  * Parses a list of function/method arguments, creates a string corresponding
  * to the arguments' list.
  */
-static char* get_args(struct parsed_symbol* sym, struct array* pmt_ref, BOOL z_term, 
+static char* get_args(struct parsed_symbol* sym, struct array* pmt_ref, BOOL z_term,
                       char open_char, char close_char)
 
 {
@@ -393,8 +393,8 @@ static char* get_args(struct parsed_symbol* sym, struct array* pmt_ref, BOOL z_t
      */
     if (z_term && *sym->current++ != 'Z') return NULL;
 
-    if (arg_collect.num == 0 || 
-        (arg_collect.num == 1 && !strcmp(arg_collect.elts[0], "void")))        
+    if (arg_collect.num == 0 ||
+        (arg_collect.num == 1 && !strcmp(arg_collect.elts[0], "void")))
         return str_printf(sym, "%cvoid%c", open_char, close_char);
     for (i = 1; i < arg_collect.num; i++)
     {
@@ -403,12 +403,12 @@ static char* get_args(struct parsed_symbol* sym, struct array* pmt_ref, BOOL z_t
 
     last = args_str ? args_str : arg_collect.elts[0];
     if (close_char == '>' && last[strlen(last) - 1] == '>')
-        args_str = str_printf(sym, "%c%s%s %c", 
+        args_str = str_printf(sym, "%c%s%s %c",
                               open_char, arg_collect.elts[0], args_str, close_char);
     else
-        args_str = str_printf(sym, "%c%s%s%c", 
+        args_str = str_printf(sym, "%c%s%s%c",
                               open_char, arg_collect.elts[0], args_str, close_char);
-    
+
     return args_str;
 }
 
@@ -516,7 +516,7 @@ static char* get_literal_string(struct parsed_symbol* sym)
         }
     } while (*++sym->current != '@');
     sym->current++;
-    if (!str_array_push(sym, ptr, sym->current - 1 - ptr, &sym->names))
+    if (!str_array_push(sym, ptr, (int)(sym->current - 1 - ptr), &sym->names))
         return NULL;
 
     return str_array_get_ref(&sym->names, sym->names.num - sym->names.start - 1);
@@ -582,7 +582,7 @@ static BOOL get_class(struct parsed_symbol* sym)
             name = str_array_get_ref(&sym->names, *sym->current++ - '0');
             break;
         case '?':
-            if (*++sym->current == '$') 
+            if (*++sym->current == '$')
             {
                 sym->current++;
                 if ((name = get_template_name(sym)) &&
@@ -613,15 +613,15 @@ static char* get_class_string(struct parsed_symbol* sym, int start)
     char*        ret;
     struct array *a = &sym->stack;
 
-    for (len = 0, i = start; i < a->num; i++)
+    for (len = 0, i = start; i < (int)a->num; i++)
     {
         assert(a->elts[i]);
-        len += 2 + strlen(a->elts[i]);
+        len += 2 + (int)strlen(a->elts[i]);
     }
     if (!(ret = und_alloc(sym, len - 1))) return NULL;
     for (len = 0, i = a->num - 1; i >= start; i--)
     {
-        sz = strlen(a->elts[i]);
+        sz = (int)strlen(a->elts[i]);
         memcpy(ret + len, a->elts[i], sz);
         len += sz;
         if (i > start)
@@ -702,7 +702,7 @@ static BOOL get_calling_convention(char ch, const char** call_conv,
 static const char* get_simple_type(char c)
 {
     const char* type_string;
-    
+
     switch (c)
     {
     case 'C': type_string = "signed char"; break;
@@ -731,7 +731,7 @@ static const char* get_simple_type(char c)
 static const char* get_extended_type(char c)
 {
     const char* type_string;
-    
+
     switch (c)
     {
     case 'D': type_string = "__int8"; break;
@@ -766,7 +766,7 @@ static BOOL demangle_datatype(struct parsed_symbol* sym, struct datatype_t* ct,
 
     assert(ct);
     ct->left = ct->right = NULL;
-    
+
     switch (dt = *sym->current++)
     {
     case '_':
@@ -791,7 +791,7 @@ static BOOL demangle_datatype(struct parsed_symbol* sym, struct datatype_t* ct,
 
             if (!(struct_name = get_class_name(sym)))
                 goto done;
-            if (!(sym->flags & UNDNAME_NO_COMPLEX_TYPE)) 
+            if (!(sym->flags & UNDNAME_NO_COMPLEX_TYPE))
             {
                 switch (dt)
                 {
@@ -839,7 +839,7 @@ static BOOL demangle_datatype(struct parsed_symbol* sym, struct datatype_t* ct,
                 unsigned                mark = sym->stack.num;
 
                 if (!get_calling_convention(*sym->current++,
-                                            &call_conv, &exported, 
+                                            &call_conv, &exported,
                                             sym->flags & ~UNDNAME_NO_ALLOCATION_LANGUAGE) ||
                     !demangle_datatype(sym, &sub_ct, pmt_ref, FALSE))
                     goto done;
@@ -848,7 +848,7 @@ static BOOL demangle_datatype(struct parsed_symbol* sym, struct datatype_t* ct,
                 if (!args) goto done;
                 sym->stack.num = mark;
 
-                ct->left  = str_printf(sym, "%s%s (%s*", 
+                ct->left  = str_printf(sym, "%s%s (%s*",
                                        sub_ct.left, sub_ct.right, call_conv);
                 ct->right = str_printf(sym, ")%s", args);
             }
@@ -944,7 +944,7 @@ static BOOL demangle_datatype(struct parsed_symbol* sym, struct datatype_t* ct,
             return FALSE;
     }
 done:
-    
+
     return ct->left != NULL;
 }
 
@@ -980,7 +980,7 @@ static BOOL handle_data(struct parsed_symbol* sym)
         case '0': access = "private: "; break;
         case '1': access = "protected: "; break;
         case '2': access = "public: "; break;
-        } 
+        }
     }
 
     if (!(sym->flags & UNDNAME_NO_MEMBER_TYPE))
@@ -1028,8 +1028,8 @@ static BOOL handle_data(struct parsed_symbol* sym)
     if (sym->flags & UNDNAME_NAME_ONLY) ct.left = ct.right = modifier = NULL;
 
     sym->result = str_printf(sym, "%s%s%s%s%s%s%s%s", access,
-                             member_type, ct.left, 
-                             modifier && ct.left ? " " : NULL, modifier, 
+                             member_type, ct.left,
+                             modifier && ct.left ? " " : NULL, modifier,
                              modifier || ct.left ? " " : NULL, name, ct.right);
     ret = TRUE;
 done:
@@ -1163,10 +1163,10 @@ static BOOL handle_method(struct parsed_symbol* sym, BOOL cast_op)
      * Yet!!! FIXME
      */
     sym->result = str_printf(sym, "%s%s%s%s%s%s%s%s%s%s%s%s",
-                             access, member_type, ct_ret.left, 
+                             access, member_type, ct_ret.left,
                              (ct_ret.left && !ct_ret.right) ? " " : NULL,
                              call_conv, call_conv ? " " : NULL, exported,
-                             name, args_str, modifier, 
+                             name, args_str, modifier,
                              modifier ? " " : NULL, ct_ret.right);
     ret = TRUE;
 done:
@@ -1446,7 +1446,7 @@ char* CDECL __unDNameEx(char* buffer, const char* mangled, int buflen,
 
     TRACE("(%p,%s,%d,%p,%p,%p,%x)\n",
           buffer, mangled, buflen, memget, memfree, unknown, flags);
-    
+
     /* The flags details is not documented by MS. However, it looks exactly
      * like the UNDNAME_ manifest constants from imagehlp.h and dbghelp.h
      * So, we copied those (on top of the file)
